requirements.txt
flask~=1.1.2
numpy~=1.18.4
boto3~=1.13.24
Pillow~=7.1.2
opencv-python~=4.2.0.34